React Native Frontend + Flask Backend
Working on implementing a React Native Frontend then connecting it with the Flask Backend, currently they are linked as the data stored within the flask temp db can be fetched via API on the frontend.

Need to look over the ports and host IPs used, as I'm currently not sure if it'll be the same for each person.

Current state of the application was tested using a physical device (android phone) using Expo Go and an Android Emulator via Android Studio