Added NotificationBell, FilterTags & Arranged Layout
Added a notification bell with pop up model and filter tag component.

I could not work with the rendered user data so I had to work with 'dummy data' for the time being.

The filter tags are currently configured to filter by budget title for now as placeholder for 'individual' & 'family' type budgets