Fixed Functionality Of FilterTags With DummyData
As suggested my team member, I accommodated a way for the filter tags to filter out the diff between individual and family budgets.

I had to make adjustments to the Backend to include 'budgetType' attribute as well as all the budget creation related functions and routes.

However I still had to work with the dummy data as my user data is not being fetched :(