Success ~
With this final change I can proudly say that Deployment Test Run #1 has been a complete success! The flask backend is deployed on Render, the frontend isn't and the web service of the backend, can safely remove the .env file that contained IP addresses.

Both mobile and web works as expected, every available aspect of the application was tested aside from any feature that deals with creation (as the models have been updated, they too need to reflect that update before being used, that's to be done on another branch though)